Add : modifier, refactor mapping structure
Replaces 4*N 2-character mappings with 4 1-character ones which get a
subsequent modifier from getchar().

Adds : as a modifier character, allowing arbitrary filename modifiers to
be used on the path.
Unfortunately, getchar() doesn't show typed characters like &showcmd
does, so entering complex : modifiers is challenging. The experience
seems to be better than the two-character mappings, when hitting a stray
character after y would grab the filename with padding and control char.

Reject . and ! commands with a count, this might be fixable later.

Also add help docs and improve the README.